+++ sydtest-mutation-driver/default.nix
@@ -4,7 +4,7 @@
 }:
 mkDerivation {
   pname = "sydtest-mutation-driver";
-  version = "0.1.0.0";
+  version = "0.2.0.0";
   src = ./.;
   isLibrary = true;
   isExecutable = true;