+++ sydtest-mutation-driver/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.2.0.0] - 2026-09-11
+
+### Fixed
+
+* The output of the run that killed a control is kept, so the flaky test behind
+  a control failure can be found.
+
+
+
 ## [0.1.0.0] - 2026-07-16
 
 * First released version.